新增管理员解决方案管理

+++ website1-module-web/src/main/java/com/fzg/model/Solutions.java
@@ -47,7 +47,7 @@ public class Solutions {
      * 0-下架 1-上架
      */
     @Schema(description = "0-下架 1-上架")
-    private Integer states;
+    private Short states;
 
     /**
      *
